Sec 5 - 1 Implement Authorization in TaskController
This commit introduces authorization checks in the show method of TaskController. Now, before a task can be viewed, the system will verify if the authenticated user has the necessary permissions to view the task. This is done using the authorize method and the TaskPolicy that has been created.

The TaskPolicy defines the authorization rules for the Task model. In this case, a user can only view a task if their id matches the user_id of the task.

Changes:

Updated show method in TaskController to include authorization check.
Created TaskPolicy to define authorization rules for Task model.